+++ include/H2DE/H2DE.h
@@ -1,4 +1,4 @@
 /**
- * Version: 2.1.4 */
-
+ * Version: 2.1.6
+ */
 #include <H2DE_engine.h>